Split iPhone complication previews into per-family files; Labs badge; review fixes
- Extracted each iPhone live-preview family into its own SwiftUI view file under
  Settings/AppleWatch/Complications/Preview (Circular/Corner/Rectangular/Inline +
  the shared rectangular gauge and the family mock), with a small
  ComplicationPreviewContext to carry the resolved styling. WatchComplicationLivePreview
  is now a thin container that owns the renderers and dispatches.
- Show a Labs badge on the Complications row in iPhone settings.
- Min/max gauge fields accept integers only and display rounded to Int.
- Address PR #5036 review: mark the on-watch diagnostics refresh() @MainActor, and
  use the view-builder overload for the corner privacy placeholder's widgetLabel.
- Reapply #5033: remove target-level CODE_SIGN_STYLE = Automatic from
  Extensions-WatchWidgets (Debug + Release), which the main merge re-introduced and
  was failing the Distribute archive.
