+++ Sources/App/Settings/AppleWatch/Complications/ComplicationListView.swift
@@ -744,409 +744,6 @@ struct WatchComplicationBuilderEditView: View {
     }
 }
 
-// MARK: - Mock family previews (Phase E)
-
-/// A small styled mock of how each WidgetKit accessory family looks, used in the builder.
-struct ComplicationFamilyPreview: View {
-    let family: WatchComplicationConfig.Family
-
-    var body: some View {
-        GeometryReader { proxy in
-            let side = min(proxy.size.width, proxy.size.height)
-            ZStack {
-                switch family {
-                case .circular:
-                    Circle().stroke(Color.accentColor, lineWidth: side * 0.12)
-                    Image(systemSymbol: .houseFill).font(.system(size: side * 0.32))
-                case .corner:
-                    Circle()
-                        .trim(from: 0.5, to: 0.75)
-                        .stroke(Color.accentColor, lineWidth: side * 0.12)
-                    Image(systemSymbol: .houseFill).font(.system(size: side * 0.22))
-                case .rectangular:
-                    RoundedRectangle(cornerRadius: side * 0.12)
-                        .stroke(Color.secondary.opacity(0.5), lineWidth: 1)
-                    VStack(alignment: .leading, spacing: side * 0.06) {
-                        Capsule().fill(Color.accentColor).frame(width: side * 0.7, height: side * 0.12)
-                        Capsule().fill(Color.secondary.opacity(0.5)).frame(width: side * 0.5, height: side * 0.1)
-                    }
-                    .padding(side * 0.12)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                case .inline:
-                    Capsule().stroke(Color.secondary.opacity(0.5), lineWidth: 1)
-                    Capsule().fill(Color.accentColor).frame(width: side * 0.6, height: side * 0.12)
-                }
-            }
-            .frame(width: side, height: side)
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-        }
-        .accessibilityHidden(true)
-    }
-}
-
-/// A horizontal progress bar with a circular value "thumb" riding the fill, and the minimum / maximum
-/// labels below the bar.
-struct RectangularGauge: View {
-    let fraction: Double
-    let minLabel: String?
-    let maxLabel: String?
-    let valueLabel: String?
-    let tint: Color
-
-    var body: some View {
-        let clamped = min(max(fraction, 0), 1)
-        VStack(spacing: 3) {
-            GeometryReader { geo in
-                let width = geo.size.width
-                ZStack(alignment: .leading) {
-                    Capsule().fill(tint.opacity(0.25)).frame(height: RectangularGaugeMetrics.barHeight)
-                    Capsule().fill(tint)
-                        .frame(
-                            width: max(RectangularGaugeMetrics.barHeight, width * clamped),
-                            height: RectangularGaugeMetrics.barHeight
-                        )
-                    if let valueLabel {
-                        RectangularGaugeThumb(value: valueLabel, tint: tint)
-                            .position(
-                                x: min(
-                                    max(width * clamped, RectangularGaugeMetrics.thumbSize / 2),
-                                    width - RectangularGaugeMetrics.thumbSize / 2
-                                ),
-                                y: RectangularGaugeMetrics.thumbSize / 2
-                            )
-                    }
-                }
-                .frame(height: RectangularGaugeMetrics.thumbSize)
-            }
-            .frame(height: RectangularGaugeMetrics.thumbSize)
-            HStack {
-                Text(verbatim: minLabel ?? " ")
-                Spacer()
-                Text(verbatim: maxLabel ?? " ")
-            }
-            .font(.system(size: 10))
-            .foregroundStyle(.secondary)
-        }
-    }
-}
-
-enum RectangularGaugeMetrics {
-    static let barHeight: CGFloat = 7
-    static let thumbSize: CGFloat = 22
-}
-
-/// Circular value marker for the rectangular progress bar: a filled disc in the bar's color with the
-/// value inside, in a contrast-aware color.
-struct RectangularGaugeThumb: View {
-    let value: String
-    let tint: Color
-
-    /// Black on light tints, white on dark ones.
-    private var contrastColor: Color {
-        var r: CGFloat = 0, g: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
-        UIColor(tint).getRed(&r, green: &g, blue: &b, alpha: &a)
-        return (0.299 * r + 0.587 * g + 0.114 * b) > 0.6 ? .black : .white
-    }
-
-    var body: some View {
-        ZStack {
-            Circle().fill(tint)
-            Text(verbatim: value)
-                .font(.system(size: 10, weight: .bold, design: .rounded))
-                .foregroundStyle(contrastColor)
-                .lineLimit(1)
-                .minimumScaleFactor(0.4)
-                .padding(1)
-        }
-        .frame(width: RectangularGaugeMetrics.thumbSize, height: RectangularGaugeMetrics.thumbSize)
-    }
-}
-
-/// A live approximation of the watch complication, mirroring `WatchWidgetsEntryView` but rendered on
-/// iPhone with current data (via Home Assistant template rendering) so the user sees the real result
-/// before saving. Not pixel-identical to watchOS, but faithful to layout, icon, value and gauge.
-struct WatchComplicationLivePreview: View {
-    let config: WatchComplicationConfig
-    /// Reports the entity's rendered unit of measurement (nil when it has none) so the editor can
-    /// decide whether to offer the "Show unit" toggle.
-    var onUnit: (String?) -> Void = { _ in }
-    @StateObject private var valueRenderer: TemplateRenderer
-    @StateObject private var gaugeRenderer: TemplateRenderer
-    @StateObject private var unitRenderer: TemplateRenderer
-
-    init(config: WatchComplicationConfig, server: Server, onUnit: @escaping (String?) -> Void = { _ in }) {
-        self.config = config
-        self.onUnit = onUnit
-        _valueRenderer = StateObject(wrappedValue: TemplateRenderer(server: server))
-        _gaugeRenderer = StateObject(wrappedValue: TemplateRenderer(server: server))
-        _unitRenderer = StateObject(wrappedValue: TemplateRenderer(server: server))
-    }
-
-    private var name: String {
-        config.name ?? config.entityDisplayName ?? config.entityId ?? ""
-    }
-
-    private var renderedValue: String {
-        if case let .success(rendered) = valueRenderer.output { return rendered }
-        return ""
-    }
-
-    private var renderedUnit: String? {
-        if case let .success(unit) = unitRenderer.output, !unit.isEmpty { return unit }
-        return nil
-    }
-
-    /// The value as displayed: appends the unit when present and enabled (entity source only).
-    private var value: String {
-        let base = renderedValue
-        guard !base.isEmpty else { return "" }
-        if config.kind == .entity, config.showsUnit(), let unit = renderedUnit {
-            return "\(base) \(unit)"
-        }
-        return base
-    }
-
-    private var fraction: Double? {
-        guard case let .success(rendered) = gaugeRenderer.output,
-              let raw = WatchComplication.percentileNumber(from: rendered) else {
-            return nil
-        }
-        return min(max(Double(raw), 0), 1)
-    }
-
-    private var showsGauge: Bool { config.showsGauge(for: config.widgetFamily) && fraction != nil }
-
-    private var range: (min: Double, max: Double)? { config.gaugeRange(for: config.widgetFamily) }
-
-    private var tint: Color {
-        config.tint(for: config.widgetFamily).map { Color(uiColor: UIColor($0)) } ?? .accentColor
-    }
-
-    /// Value/text color; defaults to white for contrast on the dark preview face.
-    private var textColor: Color {
-        config.textColor(for: config.widgetFamily).map { Color(uiColor: UIColor(hex: $0)) } ?? .white
-    }
-
-    private var showsValue: Bool { config.showsValue(for: config.widgetFamily) }
-
-    private var showsName: Bool { config.showsName(for: config.widgetFamily) }
-
-    private var showsIcon: Bool { config.showsIcon(for: config.widgetFamily) }
-
-    private var showsMin: Bool { config.showsMin(for: config.widgetFamily) }
-
-    private var showsMax: Bool { config.showsMax(for: config.widgetFamily) }
-
-    private var iconColor: Color {
-        config.iconColor.map { Color(uiColor: UIColor(hex: $0)) } ?? .white
-    }
-
-    private var iconImage: Image? {
-        guard showsIcon, let iconName = config.iconName else { return nil }
-        let image = MaterialDesignIcons(serversideValueNamed: iconName)
-            .image(ofSize: CGSize(width: 64, height: 64), color: UIColor(iconColor))
-        return Image(uiImage: image)
-    }
-
-    /// True while any of the template renderers is still evaluating.
-    private var isLoading: Bool {
-        [valueRenderer.output, gaugeRenderer.output, unitRenderer.output].contains(.loading)
-    }
-
-    var body: some View {
-        content
-            .frame(maxWidth: .infinity)
-            .frame(height: 150)
-            .overlay {
-                if isLoading {
-                    ProgressView()
-                        .controlSize(.large)
-                }
-            }
-            .onAppear(perform: applyTemplates)
-            .onChange(of: config) { _ in applyTemplates() }
-            .onChange(of: unitRenderer.output) { _ in onUnit(renderedUnit) }
-    }
-
-    @ViewBuilder
-    private var content: some View {
-        switch config.widgetFamily {
-        case .circular:
-            circular
-        case .corner:
-            corner
-        case .rectangular:
-            rectangular
-        case .inline:
-            inline
-        }
-    }
-
-    /// Corner: content tucked in the screen corner with a curved gauge along the bezel. Approximated
-    /// here with the value above a linear gauge, anchored to the bottom of the face.
-    private var corner: some View {
-        ZStack {
-            Circle().fill(Color.black)
-            VStack(spacing: 2) {
-                Spacer()
-                iconImage?
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 20, height: 20)
-                if showsValue, !value.isEmpty {
-                    Text(value)
-                        .font(.system(size: 15, weight: .semibold, design: .rounded))
-                        .foregroundStyle(textColor)
-                        .lineLimit(1)
-                        .minimumScaleFactor(0.5)
-                }
-                if showsGauge, let fraction {
-                    Gauge(value: fraction) { EmptyView() }
-                        .gaugeStyle(.accessoryLinearCapacity)
-                        .tint(tint)
-                        .frame(width: 66)
-                }
-            }
-            .padding(12)
-        }
-        .frame(width: 100, height: 100)
-        .environment(\.colorScheme, .dark)
-    }
-
-    /// Icon / value / name shown in the middle of a circular complication, each per its toggle.
-    private var centerContent: some View {
-        VStack(spacing: 0) {
-            iconImage?
-                .resizable()
-                .scaledToFit()
-                .frame(width: 20, height: 20)
-            if showsValue, !value.isEmpty {
-                Text(value)
-                    .font(.system(size: 14, weight: .semibold, design: .rounded))
-                    .foregroundStyle(textColor)
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.5)
-            }
-            if showsName, !name.isEmpty {
-                Text(name)
-                    .font(.system(size: 9))
-                    .foregroundStyle(textColor.opacity(0.7))
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.5)
-            }
-        }
-    }
-
-    /// Uses the real WidgetKit accessory gauge styles so the preview matches watchOS: an open arc
-    /// (with min/max labels) or a full capacity ring, otherwise just the icon/value.
-    @ViewBuilder
-    private var circular: some View {
-        ZStack {
-            Circle().fill(Color.black)
-            Group {
-                if showsGauge, let fraction {
-                    switch config.gaugeStyle(for: config.widgetFamily) {
-                    case .open:
-                        Gauge(value: fraction) {
-                            EmptyView()
-                        } currentValueLabel: {
-                            centerContent
-                        } minimumValueLabel: {
-                            Text(verbatim: (showsMin ? range.map { label($0.min) } : nil) ?? "")
-                        } maximumValueLabel: {
-                            Text(verbatim: (showsMax ? range.map { label($0.max) } : nil) ?? "")
-                        }
-                        .gaugeStyle(.accessoryCircular)
-                        .tint(tint)
-                    case .capacity:
-                        Gauge(value: fraction) {
-                            EmptyView()
-                        } currentValueLabel: {
-                            centerContent
-                        }
-                        .gaugeStyle(.accessoryCircularCapacity)
-                        .tint(tint)
-                    }
-                } else {
-                    centerContent
-                }
-            }
-            .padding(12)
-        }
-        .frame(width: 100, height: 100)
-        .environment(\.colorScheme, .dark)
-    }
-
-    private func label(_ value: Double) -> String {
-        String(Int(value.rounded()))
-    }
-
-    private var rectangular: some View {
-        HStack(spacing: 8) {
-            iconImage?
-                .resizable()
-                .scaledToFit()
-                .frame(width: 22, height: 22)
-            VStack(alignment: .leading, spacing: 2) {
-                if showsName {
-                    Text(name).font(.system(size: 13, weight: .semibold)).lineLimit(1).foregroundStyle(textColor)
-                }
-                if showsGauge, let fraction {
-                    RectangularGauge(
-                        fraction: fraction,
-                        minLabel: showsMin ? range.map { label($0.min) } : nil,
-                        maxLabel: showsMax ? range.map { label($0.max) } : nil,
-                        valueLabel: showsValue && !value.isEmpty ? value : nil,
-                        tint: tint
-                    )
-                } else if showsValue, !value.isEmpty {
-                    Text(value).font(.system(size: 13)).foregroundStyle(textColor.opacity(0.85)).lineLimit(1)
-                }
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(12)
-        .frame(width: 200)
-        .background(RoundedRectangle(cornerRadius: 14).fill(Color.black))
-    }
-
-    private var inline: some View {
-        // Inline has no icon or custom colors; name and value are joined with " - ".
-        Text([showsName ? name : "", showsValue ? value : ""].filter { !$0.isEmpty }.joined(separator: " - "))
-            .font(.system(size: 15)).lineLimit(1)
-            .foregroundStyle(.white)
-            .padding(.horizontal, 14).padding(.vertical, 8)
-            .background(Capsule().fill(Color.black))
-    }
-
-    private func applyTemplates() {
-        switch config.kind {
-        case .entity:
-            guard let entityId = config.entityId else {
-                valueRenderer.updateTemplate("")
-                gaugeRenderer.updateTemplate("")
-                unitRenderer.updateTemplate("")
-                return
-            }
-            valueRenderer.updateTemplate("{{ states('\(entityId)') }}")
-            unitRenderer.updateTemplate("{{ state_attr('\(entityId)', 'unit_of_measurement') or '' }}")
-            if let range = config.gaugeRange(for: config.widgetFamily) {
-                let source = config.gaugeAttribute(for: config.widgetFamily)
-                    .map { "state_attr('\(entityId)', '\($0)')" } ?? "states('\(entityId)')"
-                gaugeRenderer.updateTemplate(
-                    "{{ ((\(source) | float(0)) - \(range.min)) / \(range.max - range.min) }}"
-                )
-            } else {
-                gaugeRenderer.updateTemplate("")
-            }
-        case .customTemplate:
-            valueRenderer.updateTemplate(config.customTextTemplate ?? "")
-            gaugeRenderer.updateTemplate(config.customGaugeTemplate ?? "")
-            unitRenderer.updateTemplate("")
-        }
-    }
-}
-
 #Preview {
     NavigationView { ComplicationsRootView() }
 }
